got rid of actor code
